requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
crewai
langchain-community
duckduckgo-search
python-dotenv
